Add type checking on invocation of alias

+++ regen/regen-umbrella/regen-example/src/main/java/regen/test/project/SimpleTest.java
@@ -14,7 +14,7 @@ public class SimpleTest {
 	
 	public static void main(String[] args) {
 		@Refinement("a == 10")
-		int a = 10;
+		double a = 10;
 		@Refinement("InRange( _, a, 122)")
 		int j = getNum();
